backend/requirements.txt
Django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# banco (dev: sqlite; prod: Postgres)
psycopg2-binary==2.9.9

# util
pandas==2.2.2
python-dotenv==1.0.1
